+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### vNext
 
+- [DAV-3196] Add widget loading state
 - [DAV-3256] Add product tour snackbar
 
 ### v1.8.0